signup+login+post-creation+getPosts-OK

+++ frontend/src/main.js
@@ -4,8 +4,6 @@ import router from './router';
 import store from './store';
 import axios from 'axios';
 
-
-
 axios.defaults.baseURL = 'http://localhost:3000/api/';
 axios.defaults.withCredentials = 'true';
 
@@ -17,34 +15,20 @@ if (token) {
   // If token expired go to login
   axios.interceptors.response.use(
     (response) => {
-      return response;
+      return response
     },
     (error) => {
-      if (error.response.status === 401) {
+      const originalRequest = error.config;
+      if (error.response.status === 401 && !originalRequest._retry) {
         console.log("error 401");
         localStorage.clear("userToken");
-        //  we need to dispatch this user to null
-        store.dispatch("user", null);
-        router.push({ name: "LogIn" });
-      }
+        store.dispatch('LogOut')
+        return router.push('/login')
+      } 
       return Promise.reject(error);
     }
   );
 }
-
-
-/*axios.interceptors.response.use(undefined, function (error) {
-  if (error) {
-    
-    const originalRequest = error.config;
-    if (error.response.status === 401 && !originalRequest._retry) {
-      
-        originalRequest._retry = true;
-        store.dispatch('LogOut')
-        return router.push('/login')
-    }
-  }
-})*/
 
 const app = createApp(App)
